Move validators to a single package
