+++ internal/api/handlers/v0/edit.go
@@ -15,7 +15,6 @@ import (
 	"github.com/modelcontextprotocol/registry/internal/validators"
 	apiv1 "github.com/modelcontextprotocol/registry/pkg/api/v1"
 	"github.com/modelcontextprotocol/registry/pkg/model"
-	"github.com/modelcontextprotocol/registry/pkg/validation"
 )
 
 // EditServerInput represents the input for editing a server
@@ -56,7 +55,7 @@ func RegisterEditEndpoints(api huma.API, registry service.RegistryService, cfg *
 		}
 
 		// Validate that only allowed extension fields are present
-		if err := validation.ValidatePublishRequestExtensions(input.RawBody); err != nil {
+		if err := validators.ValidatePublishRequestExtensions(input.RawBody); err != nil {
 			return nil, huma.Error400BadRequest("Invalid request format", err)
 		}
