+++ internal/service/registry_service.go
@@ -10,9 +10,9 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/modelcontextprotocol/registry/internal/database"
+	"github.com/modelcontextprotocol/registry/internal/validators"
 	apiv1 "github.com/modelcontextprotocol/registry/pkg/api/v1"
 	"github.com/modelcontextprotocol/registry/pkg/model"
-	"github.com/modelcontextprotocol/registry/pkg/validation"
 )
 
 const maxServerVersionsPerServer = 10000
@@ -133,12 +133,12 @@ func (s *registryServiceImpl) Publish(req apiv1.PublishRequest) (*apiv1.ServerRe
 	defer cancel()
 
 	// Validate the request
-	if err := validation.ValidatePublisherExtensions(req); err != nil {
+	if err := validators.ValidatePublisherExtensions(req); err != nil {
 		return nil, err
 	}
 
 	// Validate server name exists and format
-	if _, err := validation.ParseServerName(req.Server); err != nil {
+	if _, err := validators.ParseServerName(req.Server); err != nil {
 		return nil, err
 	}
 
@@ -150,7 +150,7 @@ func (s *registryServiceImpl) Publish(req apiv1.PublishRequest) (*apiv1.ServerRe
 	}
 
 	// Validate reverse-DNS namespace matching for remote URLs
-	if err := validation.ValidateRemoteNamespaceMatch(req.Server); err != nil {
+	if err := validators.ValidateRemoteNamespaceMatch(req.Server); err != nil {
 		return nil, err
 	}
 
@@ -201,7 +201,7 @@ func (s *registryServiceImpl) Publish(req apiv1.PublishRequest) (*apiv1.ServerRe
 	}
 
 	// Extract publisher extensions from request
-	publisherExtensions := validation.ExtractPublisherExtensions(req)
+	publisherExtensions := validators.ExtractPublisherExtensions(req)
 
 	// Create registry metadata with service-determined values
 	registryMetadata := apiv1.RegistryExtensions{
@@ -261,12 +261,12 @@ func (s *registryServiceImpl) EditServer(id string, req apiv1.PublishRequest) (*
 	defer cancel()
 
 	// Validate the request
-	if err := validation.ValidatePublisherExtensions(req); err != nil {
+	if err := validators.ValidatePublisherExtensions(req); err != nil {
 		return nil, err
 	}
 
 	// Validate server name exists and format
-	if _, err := validation.ParseServerName(req.Server); err != nil {
+	if _, err := validators.ParseServerName(req.Server); err != nil {
 		return nil, err
 	}
 
@@ -278,7 +278,7 @@ func (s *registryServiceImpl) EditServer(id string, req apiv1.PublishRequest) (*
 	}
 
 	// Validate reverse-DNS namespace matching for remote URLs
-	if err := validation.ValidateRemoteNamespaceMatch(req.Server); err != nil {
+	if err := validators.ValidateRemoteNamespaceMatch(req.Server); err != nil {
 		return nil, err
 	}
 
@@ -288,7 +288,7 @@ func (s *registryServiceImpl) EditServer(id string, req apiv1.PublishRequest) (*
 	}
 
 	// Extract publisher extensions from request
-	publisherExtensions := validation.ExtractPublisherExtensions(req)
+	publisherExtensions := validators.ExtractPublisherExtensions(req)
 
 	// Update server in database
 	serverRecord, err := s.db.UpdateServer(ctx, id, req.Server, publisherExtensions)
